Download and read a GMT file from a URL

Downloads a Gene Matrix Transposed (GMT) file from a specified URL, reads it into R as a list, and removes the temporary file afterward.

Usage

getGMT(url, name = NULL, cache_dir = NULL, redownload = FALSE)

Arguments

url

A character string specifying the URL to a GMT file.

name

Optional name for the GMT file (defaults to the portion after the last '=' in the URL).

cache_dir

Optional directory in which to cache/download the GMT file.

redownload

Logical; if TRUE, forces re-download even if cached.

Value

A named list where each element is a character vector of gene names for a given gene set.

Examples

url <- paste0(
    "https://maayanlab.cloud/Enrichr/geneSetLibrary",
    "?mode=text&libraryName=KEGG_2019_Human"
)
gmt_list <- getGMT(url)
#> Auto-detected name: KEGG_2019_Human
#> Using cached file for KEGG_2019_Human
# list available gene sets
names(gmt_list)
#>   [1] "ABC transporters"                                          
#>   [2] "AGE-RAGE signaling pathway in diabetic complications"      
#>   [3] "AMPK signaling pathway"                                    
#>   [4] "Acute myeloid leukemia"                                    
#>   [5] "Adherens junction"                                         
#>   [6] "Adipocytokine signaling pathway"                           
#>   [7] "Adrenergic signaling in cardiomyocytes"                    
#>   [8] "African trypanosomiasis"                                   
#>   [9] "Alanine, aspartate and glutamate metabolism"               
#>  [10] "Alcoholism"                                                
#>  [11] "Aldosterone synthesis and secretion"                       
#>  [12] "Aldosterone-regulated sodium reabsorption"                 
#>  [13] "Allograft rejection"                                       
#>  [14] "Alzheimer disease"                                         
#>  [15] "Amino sugar and nucleotide sugar metabolism"               
#>  [16] "Aminoacyl-tRNA biosynthesis"                               
#>  [17] "Amoebiasis"                                                
#>  [18] "Amphetamine addiction"                                     
#>  [19] "Amyotrophic lateral sclerosis (ALS)"                       
#>  [20] "Antigen processing and presentation"                       
#>  [21] "Apelin signaling pathway"                                  
#>  [22] "Apoptosis"                                                 
#>  [23] "Arachidonic acid metabolism"                               
#>  [24] "Arginine and proline metabolism"                           
#>  [25] "Arginine biosynthesis"                                     
#>  [26] "Arrhythmogenic right ventricular cardiomyopathy (ARVC)"    
#>  [27] "Ascorbate and aldarate metabolism"                         
#>  [28] "Asthma"                                                    
#>  [29] "Autoimmune thyroid disease"                                
#>  [30] "Autophagy"                                                 
#>  [31] "Axon guidance"                                             
#>  [32] "B cell receptor signaling pathway"                         
#>  [33] "Bacterial invasion of epithelial cells"                    
#>  [34] "Basal cell carcinoma"                                      
#>  [35] "Basal transcription factors"                               
#>  [36] "Base excision repair"                                      
#>  [37] "Bile secretion"                                            
#>  [38] "Biosynthesis of unsaturated fatty acids"                   
#>  [39] "Bladder cancer"                                            
#>  [40] "Breast cancer"                                             
#>  [41] "Butanoate metabolism"                                      
#>  [42] "C-type lectin receptor signaling pathway"                  
#>  [43] "Caffeine metabolism"                                       
#>  [44] "Calcium signaling pathway"                                 
#>  [45] "Carbohydrate digestion and absorption"                     
#>  [46] "Cardiac muscle contraction"                                
#>  [47] "Cell adhesion molecules (CAMs)"                            
#>  [48] "Cell cycle"                                                
#>  [49] "Cellular senescence"                                       
#>  [50] "Central carbon metabolism in cancer"                       
#>  [51] "Chagas disease (American trypanosomiasis)"                 
#>  [52] "Chemical carcinogenesis"                                   
#>  [53] "Chemokine signaling pathway"                               
#>  [54] "Cholesterol metabolism"                                    
#>  [55] "Choline metabolism in cancer"                              
#>  [56] "Cholinergic synapse"                                       
#>  [57] "Chronic myeloid leukemia"                                  
#>  [58] "Circadian entrainment"                                     
#>  [59] "Circadian rhythm"                                          
#>  [60] "Citrate cycle (TCA cycle)"                                 
#>  [61] "Cocaine addiction"                                         
#>  [62] "Collecting duct acid secretion"                            
#>  [63] "Colorectal cancer"                                         
#>  [64] "Complement and coagulation cascades"                       
#>  [65] "Cortisol synthesis and secretion"                          
#>  [66] "Cushing syndrome"                                          
#>  [67] "Cysteine and methionine metabolism"                        
#>  [68] "Cytokine-cytokine receptor interaction"                    
#>  [69] "Cytosolic DNA-sensing pathway"                             
#>  [70] "D-Glutamine and D-glutamate metabolism"                    
#>  [71] "DNA replication"                                           
#>  [72] "Dilated cardiomyopathy (DCM)"                              
#>  [73] "Dopaminergic synapse"                                      
#>  [74] "Drug metabolism"                                           
#>  [75] "ECM-receptor interaction"                                  
#>  [76] "Endocrine and other factor-regulated calcium reabsorption" 
#>  [77] "Endocytosis"                                               
#>  [78] "Endometrial cancer"                                        
#>  [79] "Epithelial cell signaling in Helicobacter pylori infection"
#>  [80] "Epstein-Barr virus infection"                              
#>  [81] "ErbB signaling pathway"                                    
#>  [82] "Estrogen signaling pathway"                                
#>  [83] "Ether lipid metabolism"                                    
#>  [84] "Fanconi anemia pathway"                                    
#>  [85] "Fat digestion and absorption"                              
#>  [86] "Fatty acid biosynthesis"                                   
#>  [87] "Fatty acid degradation"                                    
#>  [88] "Fatty acid elongation"                                     
#>  [89] "Fc epsilon RI signaling pathway"                           
#>  [90] "Fc gamma R-mediated phagocytosis"                          
#>  [91] "Ferroptosis"                                               
#>  [92] "Fluid shear stress and atherosclerosis"                    
#>  [93] "Focal adhesion"                                            
#>  [94] "Folate biosynthesis"                                       
#>  [95] "FoxO signaling pathway"                                    
#>  [96] "Fructose and mannose metabolism"                           
#>  [97] "GABAergic synapse"                                         
#>  [98] "Galactose metabolism"                                      
#>  [99] "Gap junction"                                              
#> [100] "Gastric acid secretion"                                    
#> [101] "Gastric cancer"                                            
#> [102] "Glioma"                                                    
#> [103] "Glucagon signaling pathway"                                
#> [104] "Glutamatergic synapse"                                     
#> [105] "Glutathione metabolism"                                    
#> [106] "Glycerolipid metabolism"                                   
#> [107] "Glycerophospholipid metabolism"                            
#> [108] "Glycine, serine and threonine metabolism"                  
#> [109] "Glycolysis / Gluconeogenesis"                              
#> [110] "Glycosaminoglycan biosynthesis"                            
#> [111] "Glycosaminoglycan degradation"                             
#> [112] "Glycosphingolipid biosynthesis"                            
#> [113] "Glycosylphosphatidylinositol (GPI)-anchor biosynthesis"    
#> [114] "Glyoxylate and dicarboxylate metabolism"                   
#> [115] "GnRH signaling pathway"                                    
#> [116] "Graft-versus-host disease"                                 
#> [117] "HIF-1 signaling pathway"                                   
#> [118] "Hedgehog signaling pathway"                                
#> [119] "Hematopoietic cell lineage"                                
#> [120] "Hepatitis B"                                               
#> [121] "Hepatitis C"                                               
#> [122] "Hepatocellular carcinoma"                                  
#> [123] "Herpes simplex virus 1 infection"                          
#> [124] "Hippo signaling pathway"                                   
#> [125] "Histidine metabolism"                                      
#> [126] "Homologous recombination"                                  
#> [127] "Human T-cell leukemia virus 1 infection"                   
#> [128] "Human cytomegalovirus infection"                           
#> [129] "Human immunodeficiency virus 1 infection"                  
#> [130] "Human papillomavirus infection"                            
#> [131] "Huntington disease"                                        
#> [132] "Hypertrophic cardiomyopathy (HCM)"                         
#> [133] "IL-17 signaling pathway"                                   
#> [134] "Inflammatory bowel disease (IBD)"                          
#> [135] "Inflammatory mediator regulation of TRP channels"          
#> [136] "Influenza A"                                               
#> [137] "Inositol phosphate metabolism"                             
#> [138] "Insulin resistance"                                        
#> [139] "Insulin secretion"                                         
#> [140] "Insulin signaling pathway"                                 
#> [141] "Intestinal immune network for IgA production"              
#> [142] "JAK-STAT signaling pathway"                                
#> [143] "Kaposi sarcoma-associated herpesvirus infection"           
#> [144] "Legionellosis"                                             
#> [145] "Leishmaniasis"                                             
#> [146] "Leukocyte transendothelial migration"                      
#> [147] "Linoleic acid metabolism"                                  
#> [148] "Long-term depression"                                      
#> [149] "Long-term potentiation"                                    
#> [150] "Longevity regulating pathway"                              
#> [151] "Lysine degradation"                                        
#> [152] "Lysosome"                                                  
#> [153] "MAPK signaling pathway"                                    
#> [154] "Malaria"                                                   
#> [155] "Mannose type O-glycan biosynthesis"                        
#> [156] "Maturity onset diabetes of the young"                      
#> [157] "Measles"                                                   
#> [158] "Melanogenesis"                                             
#> [159] "Melanoma"                                                  
#> [160] "Metabolism of xenobiotics by cytochrome P450"              
#> [161] "MicroRNAs in cancer"                                       
#> [162] "Mineral absorption"                                        
#> [163] "Mismatch repair"                                           
#> [164] "Mitophagy"                                                 
#> [165] "Morphine addiction"                                        
#> [166] "Mucin type O-glycan biosynthesis"                          
#> [167] "N-Glycan biosynthesis"                                     
#> [168] "NF-kappa B signaling pathway"                              
#> [169] "NOD-like receptor signaling pathway"                       
#> [170] "Natural killer cell mediated cytotoxicity"                 
#> [171] "Necroptosis"                                               
#> [172] "Neomycin, kanamycin and gentamicin biosynthesis"           
#> [173] "Neuroactive ligand-receptor interaction"                   
#> [174] "Neurotrophin signaling pathway"                            
#> [175] "Nicotinate and nicotinamide metabolism"                    
#> [176] "Nicotine addiction"                                        
#> [177] "Nitrogen metabolism"                                       
#> [178] "Non-alcoholic fatty liver disease (NAFLD)"                 
#> [179] "Non-homologous end-joining"                                
#> [180] "Non-small cell lung cancer"                                
#> [181] "Notch signaling pathway"                                   
#> [182] "Nucleotide excision repair"                                
#> [183] "Olfactory transduction"                                    
#> [184] "One carbon pool by folate"                                 
#> [185] "Oocyte meiosis"                                            
#> [186] "Osteoclast differentiation"                                
#> [187] "Other glycan degradation"                                  
#> [188] "Other types of O-glycan biosynthesis"                      
#> [189] "Ovarian steroidogenesis"                                   
#> [190] "Oxidative phosphorylation"                                 
#> [191] "Oxytocin signaling pathway"                                
#> [192] "PI3K-Akt signaling pathway"                                
#> [193] "PPAR signaling pathway"                                    
#> [194] "Pancreatic cancer"                                         
#> [195] "Pancreatic secretion"                                      
#> [196] "Pantothenate and CoA biosynthesis"                         
#> [197] "Parathyroid hormone synthesis, secretion and action"       
#> [198] "Parkinson disease"                                         
#> [199] "Pathogenic Escherichia coli infection"                     
#> [200] "Pathways in cancer"                                        
#> [201] "Pentose and glucuronate interconversions"                  
#> [202] "Pentose phosphate pathway"                                 
#> [203] "Peroxisome"                                                
#> [204] "Pertussis"                                                 
#> [205] "Phagosome"                                                 
#> [206] "Phenylalanine metabolism"                                  
#> [207] "Phenylalanine, tyrosine and tryptophan biosynthesis"       
#> [208] "Phosphatidylinositol signaling system"                     
#> [209] "Phospholipase D signaling pathway"                         
#> [210] "Phosphonate and phosphinate metabolism"                    
#> [211] "Phototransduction"                                         
#> [212] "Platelet activation"                                       
#> [213] "Porphyrin and chlorophyll metabolism"                      
#> [214] "Primary bile acid biosynthesis"                            
#> [215] "Primary immunodeficiency"                                  
#> [216] "Prion diseases"                                            
#> [217] "Progesterone-mediated oocyte maturation"                   
#> [218] "Prolactin signaling pathway"                               
#> [219] "Propanoate metabolism"                                     
#> [220] "Prostate cancer"                                           
#> [221] "Proteasome"                                                
#> [222] "Protein digestion and absorption"                          
#> [223] "Protein export"                                            
#> [224] "Protein processing in endoplasmic reticulum"               
#> [225] "Proteoglycans in cancer"                                   
#> [226] "Proximal tubule bicarbonate reclamation"                   
#> [227] "Purine metabolism"                                         
#> [228] "Pyrimidine metabolism"                                     
#> [229] "Pyruvate metabolism"                                       
#> [230] "RIG-I-like receptor signaling pathway"                     
#> [231] "RNA degradation"                                           
#> [232] "RNA polymerase"                                            
#> [233] "RNA transport"                                             
#> [234] "Rap1 signaling pathway"                                    
#> [235] "Ras signaling pathway"                                     
#> [236] "Regulation of actin cytoskeleton"                          
#> [237] "Regulation of lipolysis in adipocytes"                     
#> [238] "Relaxin signaling pathway"                                 
#> [239] "Renal cell carcinoma"                                      
#> [240] "Renin secretion"                                           
#> [241] "Renin-angiotensin system"                                  
#> [242] "Retinol metabolism"                                        
#> [243] "Retrograde endocannabinoid signaling"                      
#> [244] "Rheumatoid arthritis"                                      
#> [245] "Riboflavin metabolism"                                     
#> [246] "Ribosome"                                                  
#> [247] "Ribosome biogenesis in eukaryotes"                         
#> [248] "SNARE interactions in vesicular transport"                 
#> [249] "Salivary secretion"                                        
#> [250] "Salmonella infection"                                      
#> [251] "Selenocompound metabolism"                                 
#> [252] "Serotonergic synapse"                                      
#> [253] "Shigellosis"                                               
#> [254] "Signaling pathways regulating pluripotency of stem cells"  
#> [255] "Small cell lung cancer"                                    
#> [256] "Sphingolipid metabolism"                                   
#> [257] "Sphingolipid signaling pathway"                            
#> [258] "Spliceosome"                                               
#> [259] "Staphylococcus aureus infection"                           
#> [260] "Starch and sucrose metabolism"                             
#> [261] "Steroid biosynthesis"                                      
#> [262] "Steroid hormone biosynthesis"                              
#> [263] "Sulfur metabolism"                                         
#> [264] "Sulfur relay system"                                       
#> [265] "Synaptic vesicle cycle"                                    
#> [266] "Synthesis and degradation of ketone bodies"                
#> [267] "Systemic lupus erythematosus"                              
#> [268] "T cell receptor signaling pathway"                         
#> [269] "TGF-beta signaling pathway"                                
#> [270] "TNF signaling pathway"                                     
#> [271] "Taste transduction"                                        
#> [272] "Taurine and hypotaurine metabolism"                        
#> [273] "Terpenoid backbone biosynthesis"                           
#> [274] "Th1 and Th2 cell differentiation"                          
#> [275] "Th17 cell differentiation"                                 
#> [276] "Thermogenesis"                                             
#> [277] "Thiamine metabolism"                                       
#> [278] "Thyroid cancer"                                            
#> [279] "Thyroid hormone signaling pathway"                         
#> [280] "Thyroid hormone synthesis"                                 
#> [281] "Tight junction"                                            
#> [282] "Toll-like receptor signaling pathway"                      
#> [283] "Toxoplasmosis"                                             
#> [284] "Transcriptional misregulation in cancer"                   
#> [285] "Tryptophan metabolism"                                     
#> [286] "Tuberculosis"                                              
#> [287] "Type I diabetes mellitus"                                  
#> [288] "Type II diabetes mellitus"                                 
#> [289] "Tyrosine metabolism"                                       
#> [290] "Ubiquinone and other terpenoid-quinone biosynthesis"       
#> [291] "Ubiquitin mediated proteolysis"                            
#> [292] "VEGF signaling pathway"                                    
#> [293] "Valine, leucine and isoleucine degradation"                
#> [294] "Vascular smooth muscle contraction"                        
#> [295] "Vasopressin-regulated water reabsorption"                  
#> [296] "Vibrio cholerae infection"                                 
#> [297] "Viral carcinogenesis"                                      
#> [298] "Viral myocarditis"                                         
#> [299] "Vitamin B6 metabolism"                                     
#> [300] "Vitamin digestion and absorption"                          
#> [301] "Wnt signaling pathway"                                     
#> [302] "alpha-Linolenic acid metabolism"                           
#> [303] "beta-Alanine metabolism"                                   
#> [304] "cAMP signaling pathway"                                    
#> [305] "cGMP-PKG signaling pathway"                                
#> [306] "mRNA surveillance pathway"                                 
#> [307] "mTOR signaling pathway"                                    
#> [308] "p53 signaling pathway"                                     
# inspect the first few genes in the first gene set
head(gmt_list[[1]])
#> [1] "ABCA1"  "ABCA10" "ABCA12" "ABCA13" "ABCA2"  "ABCA3"